refactored with blueprint & create_app
