+++ project/decorators.py
@@ -1,6 +1,8 @@
 import functools
+
 from flask import jsonify
 from flask_jwt_extended import get_jwt_identity
+
 from project.user.models import User
