+++ wsgi.py
@@ -1,4 +1,6 @@
-from project import app
+from project import create_app
+
+app = create_app()
 
 if __name__ == '__main__':
     app.run()
